requirements.txt
cffi==1.16.0
cryptography==41.0.7
fernet==1.0.1
pyaes==1.6.1
pycparser==2.21
pycryptodome==3.19.0
